ROCKS.BCDiag

ROCKS.accuracyplot

ROCKS.bcdiag

ROCKS.biasplot

ROCKS.concordance

ROCKS.concordance

ROCKS.cumliftable

ROCKS.cumliftcurve

ROCKS.ksplot

ROCKS.kstest

ROCKS.liftable

ROCKS.liftcurve

ROCKS.ranks

ROCKS.rocplot

ROCKS.BCDiag — Type

BCDiag

A structure of diagnostic properties of a Binary Classifier.

Facilitates summary plots and tables.

ROCKS.accuracyplot — Method

accuracyplot(x::BCDiag; util=[1, 0, 0, 1])

Using util values for [TP, FN, FP, TN], produce accuracy plot and its [max, argmax, argdep]. Default util values of [1, 0, 0, 1] gives the standard accuracy value of (TP+TN)/N.

ROCKS.bcdiag — Method

bcdiag(target, pred; groups = 100, rev = true, tie = 1e-6)

Perform diagnostics of a binary classifier. target is a 2 level categorical variable, pred is probability of class 1. groups is the number of bins to use for plotting/printing. rev = true orders pred from high to low. tie is the tolerance of pred where values are considered tied.

Returns a BCDiag struct which can be used for plotting or printing:

biasplot is calibration plot of target response rate vs. pred response rate

ksplot produces ksplot of cumulative distributions

rocplot plots the Receiver Operating Characteristics curve

accuracyplot plots the accuracy curve with adjustable utility

liftcurve is the lift curve

cumliftsurve is the cumulative lift surve

liftable is the lift table as a DataFrame

cumliftable is the cumulative lift table as a DataFrame

ROCKS.biasplot — Method

biasplot(x::BCDiag)

returns a bias calibration plot of x - actual response vs. predicted response

ROCKS.concordance — Function

concordance(class::BitVector, var::Vector, tie=1e-6)

Concordance calculation: class is a 2 level categorical target variable, var is a predictor, tie is a region where predictor values are considered as tied.

Pair-wise comparison between class 1 with class 0 values are made as follows:

class 1 value > class 0 value is Concordant

class 1 value ≈ class 0 value (within tie) is Tied

class 1 value < class 0 value is Discordant

Returns:

concordant, number of concordant comparisons

tied, number of tied comparisons

discordant, number of discordant comparisons

auroc, or C, is (Concordant + 0.5Tied) / Total comparisons; same as numeric integration of ROC curve

gini, 2C-1, also known as Somer's D, is (Concordant - Discordant) / Total comparisons

Concordance calculation is the same as numeric integration of the ROC curve, but knowledge of concordant/tied/discordant is more informative than area under curve.

Note:

Goodman-Kruskal Gamma is (Concordant - Discordant) / (Concordant + Discordant)

Kendall's Tau is (Concordant - Discordant) / (0.5 x Total count x (Total count - 1))

ROCKS.concordance — Method

concordance(class::BitVector, var::Vector, tie::Function)

Concordance calculation with a function to define tied region. More generally flexible when comparing, e.g., income of two groups, where the tied region is not constant but is a percentage of the income for instance.

ROCKS.cumliftable — Method

cumliftable(x::BCDiag)

returns a cumulative lift table of x as a DataFrame

ROCKS.cumliftcurve — Method

cumliftcurve(x::BCDiag)

returns a cumulative lift curve plot of x - cumulative actual and predicted vs. depth

ROCKS.ksplot — Method

ksplot(x::BCDiag)

returns a KS plot of x - CDF1 (True Positive) and CDF0 (False Positive) versus depth

ROCKS.kstest — Method

kstest(class, var; rev = true)

Calculate empirical 2 sample Kolmogorov-Smirnov statistic and its location. class is a 2 level categorical variable, var is the distribution to analyze.

Returns:

n, total number of observations

n1, number of observations of class 1

n0, number of observations of class 0

baserate, incidence rate of class 1

ks, the maximum separation between the two cumulative distributions

ksarg, the value of var at which maximum separation is achieved

ksdep, depth of ksarg in the sorted values of var

rev = true counts depth from high value towards low value

ROCKS.liftable — Method

liftable(x::BCDiag)

returns a lift table of x as a DataFrame

ROCKS.liftcurve — Method

liftcurve(x::BCDiag)

returns a lift curve plot of x - actual and predicted versus depth

ROCKS.ranks — Method

ranks(x; groups = 10, rank = tiedrank, rev = false)

Return a variable which bins x into groups number of bins. The rank keyword allows different ranking method; use rev = true to reverse sort so that small bin number is large value of x. Missing values are assigned to group missing.

Default values of rank = tiedrank and rev = false results in similar grouping as SAS PROC RANK groups=n tied=mean.

ROCKS.rocplot — Method

rocplot(x::BCDiag)

returns a ROC plot of x - CDF1 (True Positive) vs. CDF0 (False Positive)